+++ internal/sonos/services.go
@@ -6,6 +6,9 @@ const (
 	controlGroupRendering    = "/MediaRenderer/GroupRenderingControl/Control"
 	controlContentDirectory  = "/MediaServer/ContentDirectory/Control"
 	controlZoneGroupTopology = "/ZoneGroupTopology/Control"
+	controlMusicServices     = "/MusicServices/Control"
+	controlDeviceProperties  = "/DeviceProperties/Control"
+	controlSystemProperties  = "/SystemProperties/Control"
 	eventAVTransport         = "/MediaRenderer/AVTransport/Event"
 	eventRenderingControl    = "/MediaRenderer/RenderingControl/Event"
 	urnAVTransport           = "urn:schemas-upnp-org:service:AVTransport:1"
@@ -13,4 +16,7 @@ const (
 	urnGroupRenderingControl = "urn:schemas-upnp-org:service:GroupRenderingControl:1"
 	urnContentDirectory      = "urn:schemas-upnp-org:service:ContentDirectory:1"
 	urnZoneGroupTopology     = "urn:schemas-upnp-org:service:ZoneGroupTopology:1"
+	urnMusicServices         = "urn:schemas-upnp-org:service:MusicServices:1"
+	urnDeviceProperties      = "urn:schemas-upnp-org:service:DeviceProperties:1"
+	urnSystemProperties      = "urn:schemas-upnp-org:service:SystemProperties:1"
 )